+++ src/client/AgendaItem.jsx
@@ -12,17 +12,32 @@ import React from 'react';
 import ListItem from '@mui/material/ListItem';
 import ListItemText from '@mui/material/ListItemText';
 import { format } from 'date-fns';
-import { useTheme } from '@mui/material/styles';
 import { useTranslation } from './contexts/LocalisationContext.jsx';
 import { getColorType } from './utils/ColorUtils.jsx';
 
 export const AgendaItem = ({ agendaItem }) => {
     // get translation context
     const { t, dfnsLocale } = useTranslation();
-    // get theme
-    const theme = useTheme();
     // get color type of background color
     const colorType = getColorType(agendaItem.calendar.bgColor);
+
+    // prepare date-fns options
+    const dfnsOptions = {
+        locale: dfnsLocale,
+    };
+
+    // render line with times
+    const renderSecondary = () => {
+        // check all day event
+        if(agendaItem.allDay === true) {
+            return undefined;
+        }
+        // check multiple days
+        if(agendaItem.multipleDays === true) {
+            return `${format(agendaItem.multipleDaysStart, 'd. MMM, HH:mm', dfnsOptions)} - ${format(agendaItem.end, 'd. MMM, HH:mm', dfnsOptions)} ${t('o\'clock')}`;
+        }
+        return `${format(agendaItem.start, 'HH:mm', dfnsOptions)} - ${format(agendaItem.end, 'HH:mm', dfnsOptions)} ${t('o\'clock')}`
+    }
 
     // render jsx
     return (
@@ -44,7 +59,7 @@ export const AgendaItem = ({ agendaItem }) => {
                     },
                 }}
                 primary={`${agendaItem.title}${agendaItem.location !== null ? `, ${agendaItem.location}` : ''}${agendaItem.text !== null ? ` (${agendaItem.text})` : ''}`}
-                secondary={agendaItem.allDay === false ? `${format(agendaItem.start, 'HH:mm', dfnsLocale)} - ${format(agendaItem.end, 'HH:mm', dfnsLocale)} ${t('o\'clock')}` : undefined}
+                secondary={renderSecondary()}
             />
         </ListItem>
     );